Split Container detail page into three components
